refactor: convert project to correct maven structure #110
