+++ src/main/java/com/group26/CMV.java
@@ -1,4 +1,4 @@
-package main;
+package com.group26;
 
 
 import java.util.ArrayList;
