+++ src/main/java/com/group26/FUV.java
@@ -1,4 +1,4 @@
-package main;
+package com.group26;
 
 public class FUV {
