+++ src/main/java/com/group26/Main.java
@@ -1,4 +1,4 @@
-package main;
+package com.group26;
 
 import java.util.Random;
