+++ src/test/java/com/group26/FUVTests.java
@@ -1,10 +1,9 @@
-package junit;
+package com.group26;
 
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 import org.junit.Test;
 
-import main.FUV;
 
 public class FUVTests {
